Separated the data methods from the underlying api methods
