+++ README.md
@@ -3,6 +3,7 @@
 This repo contains a Python SDK for Cisco Viptela vManage as well as a CLI and Ansible modules that leverage the SDK.
 The previous stable release can be found at https://github.com/CiscoDevNet/python-viptela/tree/0.1.7.  The `master` branch is under active development and may not offer a stable interface.
 
+* [SDK](https://python-viptela.readthedocs.io/en/latest/)
 * [CLI](#vmanage-command-line-interface)
 * [Ansible Modules](ansible/README.md)
