+++ docs/source/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to Python viptela's documentation!
+Welcome to the Python Viptela SDK documentation!
 ==========================================
 
 `python-viptela` provides a PYthon SDK for interacting with Cisco SDWAN (formerly Viptela).
