+++ vmanage/api/device_templates.py
@@ -4,10 +4,8 @@
 import json
 import re
 from vmanage.api.feature_templates import FeatureTemplates
-from vmanage.api.device import Device
 from vmanage.api.http_methods import HttpMethods
 from vmanage.data.parse_methods import ParseMethods
-from vmanage.api.utilities import Utilities
 from vmanage.utils import list_to_dict
 
 
@@ -261,83 +259,6 @@ class DeviceTemplates(object):
             url = f"{self.base_url}template/device/feature/{device_template['templateId']}"
             response = HttpMethods(self.session, url).request('PUT', payload=json.dumps(device_template))
         return response
-
-    def import_attachment_list(self, attachment_list, check_mode=False, update=False):
-        """Import a list of device attachments to vManage.
-
-
-        Args:
-            check_mode (bool): Only check to see if changes would be made
-            update (bool): Update the template if it exists
-
-        Returns:
-            result (list): Returns the diffs of the updates.
-
-        """
-        attachment_updates = {}
-        attachment_failures = {}
-        action_id_list = []
-        device_template_dict = self.get_device_template_dict()
-        vmanage_device = Device(self.session, self.host, self.port)
-        for attachment in attachment_list:
-            if attachment['template'] in device_template_dict:
-                if attachment['device_type'] == 'vedge':
-                    # The UUID is fixes from the serial file/upload
-                    device_uuid = attachment['uuid']
-                else:
-                    # If this is not a vedge, we need to get the UUID from the vmanage since
-                    # it is generated by that vmanage
-                    device_status = vmanage_device.get_device_status(attachment['host_name'], key='host-name')
-                    if device_status:
-                        device_uuid = device_status['uuid']
-                    else:
-                        raise Exception(f"Cannot find UUID for {attachment['host_name']}")
-
-                template_id = device_template_dict[attachment['template']]['templateId']
-                attached_uuid_list = self.get_attachments(template_id, key='uuid')
-                if device_uuid in attached_uuid_list:
-                    # The device is already attached to the template.  We need to see if any of
-                    # the input changed, so we make an API call to get the input on last attach
-                    existing_template_input = self.get_template_input(
-                        device_template_dict[attachment['template']]['templateId'], [device_uuid])
-                    current_variables = existing_template_input['data'][0]
-                    changed = False
-                    for property_name in attachment['variables']:
-                        # Check to see if any of the passed in varibles have changed from what is
-                        # already on the attachment.  We are are not checking to see if the
-                        # correct variables are here.  That will be done on attachment.
-                        if ((property_name in current_variables) and
-                            (str(attachment['variables'][property_name]) != str(current_variables[property_name]))):
-                            changed = True
-                    if changed:
-                        if not check_mode and update:
-                            action_id = self.attach_to_template(template_id, device_uuid, attachment['system_ip'],
-                                                                attachment['host_name'], attachment['site_id'],
-                                                                attachment['variables'])
-                            action_id_list.append(action_id)
-                else:
-                    if not check_mode:
-                        action_id = self.attach_to_template(template_id, device_uuid, attachment['system_ip'],
-                                                            attachment['host_name'], attachment['site_id'],
-                                                            attachment['variables'])
-                        action_id_list.append(action_id)
-            else:
-                raise Exception(f"No template named Template {attachment['templateName']}")
-
-        # pp = pprint.PrettyPrinter(indent=2)
-        utilities = Utilities(self.session, self.host)
-        # Batch the waits so that the peocessing of the attachments is in parallel
-        for action_id in action_id_list:
-            result = utilities.waitfor_action_completion(action_id)
-            data = result['action_response']['data'][0]
-            # pp.pprint(data)
-            if result['action_status'] == 'failure':
-                attachment_failures.update({data['uuid']: data['currentActivity']})
-            else:
-                attachment_updates.update({data['uuid']: data['currentActivity']})
-
-        result = {'updates': attachment_updates, 'failures': attachment_failures}
-        return result
 
     def attach_to_template(self, template_id, uuid, system_ip, host_name, site_id, variables):
         """Attach and device to a template
