+++ vmanage/apps/files.py
@@ -6,12 +6,14 @@ import os
 import yaml
 from vmanage.api.device_templates import DeviceTemplates
 from vmanage.api.feature_templates import FeatureTemplates
-from vmanage.data.template_methods import TemplateMethods
+from vmanage.data.template_data import TemplateData
+from vmanage.data.policy_data import PolicyData
 from vmanage.api.policy_lists import PolicyLists
 from vmanage.api.policy_definitions import PolicyDefinitions
 from vmanage.api.local_policy import LocalPolicy
 from vmanage.api.central_policy import CentralPolicy
 from vmanage.api.device import Device
+from vmanage.utils import list_to_dict
 
 
 class Files(object):
@@ -36,23 +38,27 @@ class Files(object):
         self.host = host
         self.port = port
         self.base_url = f'https://{self.host}:{self.port}/dataservice/'
+        self.device_templates = DeviceTemplates(self.session, self.host, self.port)
+        self.feature_templates = FeatureTemplates(self.session, self.host, self.port)
+        self.template_data = TemplateData(self.session, self.host, self.port)
+        self.policy_data = PolicyData(self.session, self.host, self.port)
+        self.policy_lists = PolicyLists(self.session, self.host, self.port)
+        self.policy_definitions = PolicyDefinitions(self.session, self.host, self.port)
+        self.local_policy = LocalPolicy(self.session, self.host, self.port)
+        self.central_policy = CentralPolicy(self.session, self.host, self.port)
+        self.vmanage_device = Device(self.session, self.host, self.port)
 
     def export_templates_to_file(self, export_file, name_list=None, template_type=None):
-
-        device_templates = DeviceTemplates(self.session, self.host, self.port)
-        feature_templates = FeatureTemplates(self.session, self.host, self.port)
-        template_methods = TemplateMethods(self.session, self.host, self.port)
-
         template_export = {}
         #pylint: disable=too-many-nested-blocks
         if template_type != 'feature':
             # Export the device templates and associated feature templates
-            device_template_list = device_templates.get_device_template_list(name_list=name_list)
+            device_template_list = self.device_templates.get_device_template_list(name_list=name_list)
             template_export.update({'vmanage_device_templates': device_template_list})
             feature_name_list = []
 
             for device_template in device_template_list:
-                device_template = template_methods.convert_device_template_to_name(device_template)
+                device_template = self.template_data.convert_device_template_to_name(device_template)
                 if name_list and 'generalTemplates' in device_template:
                     # If a name_list is specified, gather a list of all the feature
                     # templates used so we know which ones we need to export
@@ -67,8 +73,7 @@ class Files(object):
                 if feature_name_list:
                     feature_name_list = list(set(feature_name_list))
         # Since device templates depend on feature templates, we always add them.
-        feature_templates = FeatureTemplates(self.session, self.host, self.port)
-        feature_template_list = feature_templates.get_feature_template_list(name_list=feature_name_list)
+        feature_template_list = self.feature_templates.get_feature_template_list(name_list=feature_name_list)
         template_export.update({'vmanage_feature_templates': feature_template_list})
 
         if export_file.endswith('.json'):
@@ -82,10 +87,6 @@ class Files(object):
 
     #pylint: disable=unused-argument
     def import_templates_from_file(self, file, update=False, check_mode=False, name_list=None, template_type=None):
-
-        template_methods = TemplateMethods(self.session, self.host, self.port)
-        vmanage_feature_templates = FeatureTemplates(self.session, self.host, self.port)
-
         feature_template_updates = []
         device_template_updates = []
         template_data = {}
@@ -130,9 +131,9 @@ class Files(object):
         # Since device templates depend on feature templates, we always add them.
         if name_list:
             pruned_feature_template_list = []
-            imported_feature_template_dict = self.list_to_dict(imported_feature_template_list,
-                                                               key_name='templateName',
-                                                               remove_key=False)
+            imported_feature_template_dict = list_to_dict(imported_feature_template_list,
+                                                          key_name='templateName',
+                                                          remove_key=False)
             for feature_template_name in name_list:
                 if feature_template_name in imported_feature_template_dict:
                     pruned_feature_template_list.append(imported_feature_template_dict[feature_template_name])
@@ -140,13 +141,14 @@ class Files(object):
             imported_feature_template_list = pruned_feature_template_list
 
         # Process the feature templates
-        feature_template_updates = vmanage_feature_templates.import_feature_template_list(
-            imported_feature_template_list, check_mode=check_mode, update=update)
+        feature_template_updates = self.template_data.import_feature_template_list(imported_feature_template_list,
+                                                                                   check_mode=check_mode,
+                                                                                   update=update)
 
         # Process the device templates
-        device_template_updates = template_methods.import_device_template_list(imported_device_template_list,
-                                                                               check_mode=check_mode,
-                                                                               update=update)
+        device_template_updates = self.template_data.import_device_template_list(imported_device_template_list,
+                                                                                 check_mode=check_mode,
+                                                                                 update=update)
 
         return {
             'feature_template_updates': feature_template_updates,
@@ -158,15 +160,10 @@ class Files(object):
     #
     def export_policy_to_file(self, export_file):
 
-        policy_lists = PolicyLists(self.session, self.host, self.port)
-        policy_definitions = PolicyDefinitions(self.session, self.host, self.port)
-        local_policy = LocalPolicy(self.session, self.host, self.port)
-        central_policy = CentralPolicy(self.session, self.host, self.port)
-
-        policy_lists_list = policy_lists.get_policy_list_list()
-        policy_definitions_list = policy_definitions.get_policy_definition_list()
-        central_policies_list = central_policy.get_central_policy_list()
-        local_policies_list = local_policy.get_local_policy_list()
+        policy_lists_list = self.policy_lists.get_policy_list_list()
+        policy_definitions_list = self.policy_data.export_policy_definition_list()
+        central_policies_list = self.policy_data.export_central_policy_list()
+        local_policies_list = self.local_policy.get_local_policy_list()
 
         policy_export = {
             'vmanage_policy_lists': policy_lists_list,
@@ -185,11 +182,6 @@ class Files(object):
             raise Exception("File format not supported")
 
     def import_policy_from_file(self, file, update=False, check_mode=False, push=False):
-
-        vmanage_policy_lists = PolicyLists(self.session, self.host, self.port)
-        vmanage_policy_definitions = PolicyDefinitions(self.session, self.host, self.port)
-        vmanage_central_policy = CentralPolicy(self.session, self.host, self.port)
-        vmanage_local_policy = LocalPolicy(self.session, self.host, self.port)
 
         policy_list_updates = []
         policy_definition_updates = []
@@ -223,25 +215,25 @@ class Files(object):
         else:
             local_policy_data = []
 
-        policy_list_updates = vmanage_policy_lists.import_policy_list_list(policy_list_data,
-                                                                           check_mode=check_mode,
-                                                                           update=update,
-                                                                           push=push)
+        policy_list_updates = self.policy_data.import_policy_list_list(policy_list_data,
+                                                                       check_mode=check_mode,
+                                                                       update=update,
+                                                                       push=push)
 
-        vmanage_policy_lists.clear_policy_list_cache()
+        self.policy_lists.clear_policy_list_cache()
 
-        policy_definition_updates = vmanage_policy_definitions.import_policy_definition_list(policy_definition_data,
-                                                                                             check_mode=check_mode,
-                                                                                             update=update,
-                                                                                             push=push)
-        central_policy_updates = vmanage_central_policy.import_central_policy_list(central_policy_data,
+        policy_definition_updates = self.policy_data.import_policy_definition_list(policy_definition_data,
                                                                                    check_mode=check_mode,
                                                                                    update=update,
                                                                                    push=push)
-        local_policy_updates = vmanage_local_policy.import_local_policy_list(local_policy_data,
+        central_policy_updates = self.policy_data.import_central_policy_list(central_policy_data,
                                                                              check_mode=check_mode,
                                                                              update=update,
                                                                              push=push)
+        local_policy_updates = self.policy_data.import_local_policy_list(local_policy_data,
+                                                                         check_mode=check_mode,
+                                                                         update=update,
+                                                                         push=push)
 
         return {
             'policy_list_updates': policy_list_updates,
@@ -251,21 +243,18 @@ class Files(object):
         }
 
     def export_attachments_to_file(self, export_file, name_list=None, device_type=None):
-        device_templates = DeviceTemplates(self.session, self.host, self.port)
-        vmanage_device = Device(self.session, self.host, self.port)
-
         if name_list is None:
             name_list = []
 
-        device_template_dict = device_templates.get_device_template_dict()
+        device_template_dict = self.device_templates.get_device_template_dict()
 
         attachments_list = []
         # Create a device config of the right type of things
         device_list = []
         if device_type in (None, 'controllers'):
-            device_list = vmanage_device.get_device_config_list('controllers')
+            device_list = self.vmanage_device.get_device_config_list('controllers')
         if device_type in (None, 'vedges'):
-            edge_list = vmanage_device.get_device_config_list('vedges')
+            edge_list = self.vmanage_device.get_device_config_list('vedges')
             device_list = device_list + edge_list
 
         for device_config in device_list:
@@ -278,8 +267,8 @@ class Files(object):
                     raise Exception(f"Could not find ID for template {device_config['template']}")
                 if name_list == [] or device_config['host-name'] in name_list:
                     variable_dict = {}
-                    template_input = device_templates.get_template_input(template_id,
-                                                                         device_id_list=[device_config['uuid']])
+                    template_input = self.device_templates.get_template_input(template_id,
+                                                                              device_id_list=[device_config['uuid']])
                     data = template_input['data'][0]
                     for column in template_input['columns']:
                         variable_dict[column['variable']] = data[column['property']]
@@ -306,8 +295,6 @@ class Files(object):
         return (len(attachments_list))
 
     def import_attachments_from_file(self, file, update=False, check_mode=False, name_list=None, template_type=None):
-        vmanage_device_templates = DeviceTemplates(self.session, self.host, self.port)
-
         template_data = {}
         # Read in the datafile
         if not os.path.exists(file):
@@ -324,7 +311,7 @@ class Files(object):
             imported_attachment_list = []
 
         # Process the device templates
-        result = vmanage_device_templates.import_attachment_list(imported_attachment_list,
-                                                                 check_mode=check_mode,
-                                                                 update=update)
+        result = self.template_data.import_attachment_list(imported_attachment_list,
+                                                           check_mode=check_mode,
+                                                           update=update)
         return result
