+++ vmanage/data/template_data.py
@@ -4,10 +4,12 @@
 import dictdiffer
 from vmanage.api.feature_templates import FeatureTemplates
 from vmanage.api.device_templates import DeviceTemplates
+from vmanage.api.utilities import Utilities
+from vmanage.api.device import Device
 from vmanage.api.local_policy import LocalPolicy
 
 
-class TemplateMethods(object):
+class TemplateData(object):
     """vManage Device Methods
 
     Responsible vManage Device Templates.
@@ -131,6 +133,72 @@ class TemplateMethods(object):
 
         return converted_generalTemplates
 
+    def import_feature_template_list(self, feature_template_list, check_mode=False, update=False):
+        """Add a list of feature templates to vManage.
+
+
+        Args:
+            check_mode (bool): Only check to see if changes would be made
+            update (bool): Update the template if it exists
+
+        Returns:
+            result (list): Returns the diffs of the updates.
+
+        """
+        # Process the feature templates
+        feature_template_updates = []
+        feature_template_dict = self.feature_templates.get_feature_template_dict(factory_default=True, remove_key=False)
+        for feature_template in feature_template_list:
+            if feature_template['templateName'] in feature_template_dict:
+                existing_template = feature_template_dict[feature_template['templateName']]
+                feature_template['templateId'] = existing_template['templateId']
+                diff = list(
+                    dictdiffer.diff(existing_template['templateDefinition'], feature_template['templateDefinition']))
+                if len(diff):
+                    feature_template_updates.append({'name': feature_template['templateName'], 'diff': diff})
+                    if not check_mode and update:
+                        self.feature_templates.update_feature_template(feature_template)
+            else:
+                diff = list(dictdiffer.diff({}, feature_template['templateDefinition']))
+                feature_template_updates.append({'name': feature_template['templateName'], 'diff': diff})
+                if not check_mode:
+                    self.feature_templates.add_feature_template(feature_template)
+
+        return feature_template_updates
+
+    def export_device_template_list(self, factory_default=False, name_list=None):
+        """Export the list of device templates.
+
+        Args:
+            factory_default (bool): Include factory default
+            name_list (list of strings): A list of template names to retreive.
+
+        Returns:
+            result (dict): All data associated with a response.
+        """
+        if name_list is None:
+            name_list = []
+        device_template_list = self.device_templates.get_device_templates()
+        return_list = []
+
+        #pylint: disable=too-many-nested-blocks
+        for device_template in device_template_list:
+            # If there is a list of template name, only return the ones asked for.
+            # Otherwise, return them all
+            if name_list and device_template['templateName'] not in name_list:
+                continue
+            obj = self.device_templates.get_device_template_object(device_template['templateId'])
+            if obj:
+                if not factory_default and obj['factoryDefault']:
+                    continue
+                obj['templateId'] = device_template['templateId']
+
+                # obj['attached_devices'] = self.get_template_attachments(device['templateId'])
+                # obj['input'] = self.get_template_input(device['templateId'])
+                converted_device_template = self.template_data.convert_device_template_to_name(obj)
+                return_list.append(converted_device_template)
+        return return_list
+
     def import_device_template_list(self, device_template_list, check_mode=False, update=False):
         """Add a list of feature templates to vManage.
 
@@ -177,3 +245,80 @@ class TemplateMethods(object):
                     self.add_device_template(converted_device_template)
 
         return device_template_updates
+
+    def import_attachment_list(self, attachment_list, check_mode=False, update=False):
+        """Import a list of device attachments to vManage.
+
+
+        Args:
+            check_mode (bool): Only check to see if changes would be made
+            update (bool): Update the template if it exists
+
+        Returns:
+            result (list): Returns the diffs of the updates.
+
+        """
+        attachment_updates = {}
+        attachment_failures = {}
+        action_id_list = []
+        device_template_dict = self.device_templates.get_device_template_dict()
+        vmanage_device = Device(self.session, self.host, self.port)
+        for attachment in attachment_list:
+            if attachment['template'] in device_template_dict:
+                if attachment['device_type'] == 'vedge':
+                    # The UUID is fixes from the serial file/upload
+                    device_uuid = attachment['uuid']
+                else:
+                    # If this is not a vedge, we need to get the UUID from the vmanage since
+                    # it is generated by that vmanage
+                    device_status = vmanage_device.get_device_status(attachment['host_name'], key='host-name')
+                    if device_status:
+                        device_uuid = device_status['uuid']
+                    else:
+                        raise Exception(f"Cannot find UUID for {attachment['host_name']}")
+
+                template_id = device_template_dict[attachment['template']]['templateId']
+                attached_uuid_list = self.device_templates.get_attachments(template_id, key='uuid')
+                if device_uuid in attached_uuid_list:
+                    # The device is already attached to the template.  We need to see if any of
+                    # the input changed, so we make an API call to get the input on last attach
+                    existing_template_input = self.device_templates.get_template_input(
+                        device_template_dict[attachment['template']]['templateId'], [device_uuid])
+                    current_variables = existing_template_input['data'][0]
+                    changed = False
+                    for property_name in attachment['variables']:
+                        # Check to see if any of the passed in varibles have changed from what is
+                        # already on the attachment.  We are are not checking to see if the
+                        # correct variables are here.  That will be done on attachment.
+                        if ((property_name in current_variables) and
+                            (str(attachment['variables'][property_name]) != str(current_variables[property_name]))):
+                            changed = True
+                    if changed:
+                        if not check_mode and update:
+                            action_id = self.attach_to_template(template_id, device_uuid, attachment['system_ip'],
+                                                                attachment['host_name'], attachment['site_id'],
+                                                                attachment['variables'])
+                            action_id_list.append(action_id)
+                else:
+                    if not check_mode:
+                        action_id = self.attach_to_template(template_id, device_uuid, attachment['system_ip'],
+                                                            attachment['host_name'], attachment['site_id'],
+                                                            attachment['variables'])
+                        action_id_list.append(action_id)
+            else:
+                raise Exception(f"No template named Template {attachment['templateName']}")
+
+        # pp = pprint.PrettyPrinter(indent=2)
+        utilities = Utilities(self.session, self.host)
+        # Batch the waits so that the peocessing of the attachments is in parallel
+        for action_id in action_id_list:
+            result = utilities.waitfor_action_completion(action_id)
+            data = result['action_response']['data'][0]
+            # pp.pprint(data)
+            if result['action_status'] == 'failure':
+                attachment_failures.update({data['uuid']: data['currentActivity']})
+            else:
+                attachment_updates.update({data['uuid']: data['currentActivity']})
+
+        result = {'updates': attachment_updates, 'failures': attachment_failures}
+        return result
